# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Acestream Scraper API",
    "description": "API for scraping and managing Acestream channels",
    "version": "2.0.0"
  },
  "paths": {
    "/api/v1/channels/": {
      "get": {
        "tags": ["channels"],
        "summary": "Get Channels",
        "description": "Get all channels with optional filtering.\nSupports both skip/limit and page/page_size parameters.",
        "operationId": "get_channels_api_v1_channels__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Skip" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "title": "Limit" }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "title": "Page" }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "title": "Page Size" }
          },
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Active Only"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Search"
            }
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Group"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ChannelResponse" },
                  "title": "Response Get Channels Api V1 Channels  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["channels"],
        "summary": "Create Channel",
        "description": "Create a new Acestream channel. If a channel with the same ID exists, update it.",
        "operationId": "create_channel_api_v1_channels__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChannelCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChannelResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/channels/{channel_id}": {
      "get": {
        "tags": ["channels"],
        "summary": "Get Channel",
        "description": "Get a specific channel by ID.",
        "operationId": "get_channel_api_v1_channels__channel_id__get",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Channel Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChannelResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["channels"],
        "summary": "Update Channel",
        "description": "Update an existing channel.",
        "operationId": "update_channel_api_v1_channels__channel_id__put",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Channel Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChannelUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChannelResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["channels"],
        "summary": "Delete Channel",
        "description": "Delete a channel.",
        "operationId": "delete_channel_api_v1_channels__channel_id__delete",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Channel Id" }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tv-channels/": {
      "get": {
        "tags": ["tv-channels"],
        "summary": "Get Tv Channels",
        "description": "Get all TV channels.",
        "operationId": "get_tv_channels_api_v1_tv_channels__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Skip" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/TVChannelResponse" },
                  "title": "Response Get Tv Channels Api V1 Tv Channels  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["tv-channels"],
        "summary": "Create Tv Channel",
        "description": "Create a new TV channel.",
        "operationId": "create_tv_channel_api_v1_tv_channels__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TVChannelCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TVChannelResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tv-channels/{tv_channel_id}": {
      "get": {
        "tags": ["tv-channels"],
        "summary": "Get Tv Channel",
        "description": "Get a specific TV channel by ID.",
        "operationId": "get_tv_channel_api_v1_tv_channels__tv_channel_id__get",
        "parameters": [
          {
            "name": "tv_channel_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Tv Channel Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TVChannelResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["tv-channels"],
        "summary": "Update Tv Channel",
        "description": "Update an existing TV channel.",
        "operationId": "update_tv_channel_api_v1_tv_channels__tv_channel_id__put",
        "parameters": [
          {
            "name": "tv_channel_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Tv Channel Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TVChannelUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TVChannelResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["tv-channels"],
        "summary": "Delete Tv Channel",
        "description": "Delete a TV channel.",
        "operationId": "delete_tv_channel_api_v1_tv_channels__tv_channel_id__delete",
        "parameters": [
          {
            "name": "tv_channel_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Tv Channel Id" }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tv-channels/{tv_channel_id}/acestreams": {
      "get": {
        "tags": ["tv-channels"],
        "summary": "Get Tv Channel Acestreams",
        "description": "Get all acestream channels associated with a TV channel.",
        "operationId": "get_tv_channel_acestreams_api_v1_tv_channels__tv_channel_id__acestreams_get",
        "parameters": [
          {
            "name": "tv_channel_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Tv Channel Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ChannelResponse" },
                  "title": "Response Get Tv Channel Acestreams Api V1 Tv Channels  Tv Channel Id  Acestreams Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["tv-channels"],
        "summary": "Associate Acestream",
        "description": "Associate an acestream channel with a TV channel.",
        "operationId": "associate_acestream_api_v1_tv_channels__tv_channel_id__acestreams_post",
        "parameters": [
          {
            "name": "tv_channel_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Tv Channel Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Association"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tv-channels/{tv_channel_id}/acestreams/{acestream_id}": {
      "delete": {
        "tags": ["tv-channels"],
        "summary": "Remove Acestream Association",
        "description": "Remove association between an acestream channel and a TV channel.",
        "operationId": "remove_acestream_association_api_v1_tv_channels__tv_channel_id__acestreams__acestream_id__delete",
        "parameters": [
          {
            "name": "tv_channel_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Tv Channel Id" }
          },
          {
            "name": "acestream_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Acestream Id" }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tv-channels/batch-assign": {
      "post": {
        "tags": ["tv-channels"],
        "summary": "Batch Assign Acestreams",
        "description": "Batch assign acestream channels to TV channels.\n\nThe request body should be a dictionary where keys are TV channel IDs\nand values are lists of acestream channel IDs.\n\nExample:\n{\n    \"1\": [\"acestream1\", \"acestream2\"],\n    \"2\": [\"acestream3\"]\n}",
        "operationId": "batch_assign_acestreams_api_v1_tv_channels_batch_assign_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Assignment Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tv-channels/bulk-update-epg": {
      "post": {
        "tags": ["tv-channels"],
        "summary": "Bulk Update Epg",
        "description": "Update EPG IDs for multiple TV channels.\n\nThis endpoint updates EPG IDs for multiple TV channels in a batch.\n\nRequest format:\n{\n    \"updates\": [\n        {\"tv_channel_id\": 1, \"epg_id\": \"new.epg.id1\"},\n        {\"tv_channel_id\": 2, \"epg_id\": \"new.epg.id2\"}\n    ]\n}",
        "operationId": "bulk_update_epg_api_v1_tv_channels_bulk_update_epg_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Update Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/epg/channels": {
      "get": {
        "tags": ["epg"],
        "summary": "Get Epg Channels",
        "description": "Get EPG channels, optionally filtered by source ID",
        "operationId": "get_epg_channels_api_v1_epg_channels_get",
        "parameters": [
          {
            "name": "source_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "title": "Source Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Skip" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EPGChannelResponse"
                  },
                  "title": "Response Get Epg Channels Api V1 Epg Channels Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/epg/channels/{channel_id}": {
      "get": {
        "tags": ["epg"],
        "summary": "Get Epg Channel",
        "description": "Get a specific EPG channel by ID",
        "operationId": "get_epg_channel_api_v1_epg_channels__channel_id__get",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Channel Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EPGChannelResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChannelCreate": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "name": { "type": "string", "title": "Name" },
          "source_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Source Url"
          },
          "group": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Group"
          },
          "logo": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Logo"
          },
          "tvg_id": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Tvg Id"
          },
          "tvg_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Tvg Name"
          },
          "is_online": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Is Online",
            "default": true
          }
        },
        "type": "object",
        "required": ["id", "name"],
        "title": "ChannelCreate",
        "description": "Schema for channel creation"
      },
      "ChannelResponse": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "name": { "type": "string", "title": "Name" },
          "source_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Source Url"
          },
          "group": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Group"
          },
          "logo": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Logo"
          },
          "tvg_id": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Tvg Id"
          },
          "tvg_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Tvg Name"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen"
          },
          "is_active": { "type": "boolean", "title": "Is Active" },
          "is_online": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Is Online"
          },
          "last_checked": {
            "anyOf": [
              { "type": "string", "format": "date-time" },
              { "type": "null" }
            ],
            "title": "Last Checked"
          },
          "check_error": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Check Error"
          },
          "tv_channel_id": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Tv Channel Id"
          },
          "epg_update_protected": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Epg Update Protected",
            "default": false
          }
        },
        "type": "object",
        "required": ["id", "name", "last_seen", "is_active"],
        "title": "ChannelResponse",
        "description": "Schema for channel response"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}